Move the provision stuff to -common
